+++ src/clusters/raft/mod.rs
@@ -1,5 +1,5 @@
 pub(crate) mod actor;
 pub(crate) mod log;
 pub(crate) mod messages;
-pub(crate) mod state;
+pub(crate) mod raft;
 pub(crate) mod transport;
